add --dry-run and --verbose

+++ conda/initialize.py
@@ -14,7 +14,8 @@ from tempfile import NamedTemporaryFile
 
 from . import CONDA_PACKAGE_ROOT
 from ._vendor.auxlib.ish import dals
-from .common.compat import PY2, on_mac, on_win, open, ensure_binary, ensure_unicode
+from .base.context import context
+from .common.compat import PY2, ensure_binary, ensure_unicode, on_mac, on_win, open
 from .common.path import (expand, get_python_short_path, get_python_site_packages_short_path,
                           win_path_ok)
 from .gateways.disk.create import copy, mkdir_p
@@ -53,7 +54,8 @@ class Result:
 def install(conda_prefix):
     plan = make_install_plan(conda_prefix)
     run_plan(plan)
-    assert not any(step['result'] == Result.NEEDS_SUDO for step in plan)
+    if not context.dry_run:
+        assert not any(step['result'] == Result.NEEDS_SUDO for step in plan)
     print_plan_results(plan)
 
 
@@ -62,12 +64,14 @@ def initialize(conda_prefix, shells, for_user, for_system, desktop_prompt):
     if os.getenv('CONDA_PIP_UNINITIALIZED') == 'true':
         plan1 = make_install_plan(conda_prefix)
         run_plan(plan1)
-        run_plan_elevated(plan1)
-        # TODO: make sure this all succeeded
+        if not context.dry_run:
+            run_plan_elevated(plan1)
+            # TODO: make sure this all succeeded
 
     plan2 = make_initialize_plan(conda_prefix, shells, for_user, for_system, desktop_prompt)
     run_plan(plan2)
-    run_plan_elevated(plan2)
+    if not context.dry_run:
+        run_plan_elevated(plan2)
     print_plan_results(plan1 + plan2)
 
 
@@ -116,6 +120,10 @@ def initialize_dev(shell, dev_env_prefix=None, conda_source_root=None):
     })
 
     run_plan(plan)
+
+    if context.dry_run:
+        print_plan_results(plan, sys.stderr)
+
     if any(step['result'] == Result.NEEDS_SUDO for step in plan):
         print("Operation failed.", file=sys.stderr)
         return 1
@@ -289,6 +297,8 @@ def make_initialize_plan(conda_prefix, shells, for_user, for_system, desktop_pro
             plan.append({
                 'function': init_cmd_exe_user.__name__,
                 'kwargs': {
+                    'target_path': 'HKEY_CURRENT_USER\\Software\\Microsoft\\'
+                                   'Command Processor\\AutoRun',
                     'conda_prefix': conda_prefix,
                 },
             })
@@ -387,9 +397,9 @@ def run_plan_from_temp_file(temp_path):
 
 def print_plan_results(plan, stream=sys.stdout):
     for step in plan:
-        print("%s\n  %s\n" % (step['kwargs']['target_path'], step['result']), file=stream)
+        print("%s\n  %s\n" % (step['kwargs']['target_path'], step.get('result')), file=stream)
 
-    changed = any(step['result'] == Result.MODIFIED for step in plan)
+    changed = any(step.get('result') == Result.MODIFIED for step in plan)
     if changed:
         print("\n==> For changes to take effect, close and re-open your current shell. <==\n",
               file=stream)
@@ -436,11 +446,16 @@ def make_entry_point(target_path, conda_prefix, module, func):
     }
 
     if new_ep_content != original_ep_content:
-        mkdir_p(dirname(conda_ep_path))
-        with open(conda_ep_path, 'w') as fdst:
-            fdst.write(new_ep_content)
-        if not on_win:
-            make_executable(conda_ep_path)
+        if not context.dry_run:
+            mkdir_p(dirname(conda_ep_path))
+            with open(conda_ep_path, 'w') as fdst:
+                fdst.write(new_ep_content)
+            if not on_win:
+                make_executable(conda_ep_path)
+        elif context.verbosity:
+            print(target_path)
+            print(make_diff(original_ep_content, new_ep_content))
+
         return Result.MODIFIED
     else:
         return Result.NO_CHANGE
@@ -455,11 +470,12 @@ def make_entry_point_exe(target_path, conda_prefix):
         if compute_md5sum(exe_path) == compute_md5sum(source_exe_path):
             return Result.NO_CHANGE
 
-    if not isdir(dirname(exe_path)):
-        mkdir_p(dirname(exe_path))
-    # prefer copy() over create_hard_link_or_copy() because of windows file deletion issues
-    # with open processes
-    copy(source_exe_path, exe_path)
+    if not context.dry_run:
+        if not isdir(dirname(exe_path)):
+            mkdir_p(dirname(exe_path))
+        # prefer copy() over create_hard_link_or_copy() because of windows file deletion issues
+        # with open processes
+        copy(source_exe_path, exe_path)
     return Result.MODIFIED
 
 
@@ -482,14 +498,15 @@ def install_conda_shortcut(target_path, conda_prefix):
     # The API for the call to 'create_shortcut' has 3
     # required arguments (path, description, filename)
     # and 4 optional ones (args, working_dir, icon_path, icon_index).
-    create_shortcut(
-        "%windir%\\System32\\cmd.exe",
-        "Conda Prompt",
-        '' + target_path,
-        ' '.join(args),
-        '' + expanduser('~'),
-        '' + icon_path,
-    )
+    if not context.dry_run:
+        create_shortcut(
+            "%windir%\\System32\\cmd.exe",
+            "Conda Prompt",
+            '' + target_path,
+            ' '.join(args),
+            '' + expanduser('~'),
+            '' + icon_path,
+        )
 
 
 def _install_file(target_path, file_content):
@@ -502,9 +519,13 @@ def _install_file(target_path, file_content):
     new_content = file_content
 
     if new_content != original_content:
-        mkdir_p(dirname(target_path))
-        with open(target_path, 'w') as fdst:
-            fdst.write(new_content)
+        if not context.dry_run:
+            mkdir_p(dirname(target_path))
+            with open(target_path, 'w') as fdst:
+                fdst.write(new_content)
+        elif context.verbosity:
+            print(target_path)
+            print(make_diff(original_content, new_content))
         return Result.MODIFIED
     else:
         return Result.NO_CHANGE
@@ -601,8 +622,13 @@ def init_sh_user(target_path, conda_prefix, shell):
         rc_content += '\n%s\n' % conda_initialize_content
 
     if rc_content != rc_original_content:
-        with open(user_rc_path, 'w') as fh:
-            fh.write(rc_content)
+        if not context.dry_run:
+            with open(user_rc_path, 'w') as fh:
+                fh.write(rc_content)
+        elif context.verbosity:
+            print(target_path)
+            print(make_diff(rc_original_content, rc_content))
+
         return Result.MODIFIED
     else:
         return Result.NO_CHANGE
@@ -619,17 +645,21 @@ def init_sh_system(target_path, conda_prefix):
         conda_sh_system_contents = ""
     conda_sh_contents = 'eval "$(\'%s\' shell.posix hook)"\n' % _conda_exe(conda_prefix)
     if conda_sh_system_contents != conda_sh_contents:
-        if lexists(conda_sh_system_path):
-            rm_rf(conda_sh_system_path)
-        mkdir_p(dirname(conda_sh_system_path))
-        with open(conda_sh_system_path, 'w') as fh:
-            fh.write(conda_sh_contents)
+        if not context.dry_run:
+            if lexists(conda_sh_system_path):
+                rm_rf(conda_sh_system_path)
+            mkdir_p(dirname(conda_sh_system_path))
+            with open(conda_sh_system_path, 'w') as fh:
+                fh.write(conda_sh_contents)
+        elif context.verbosity:
+            print(target_path)
+            print(make_diff(conda_sh_contents, conda_sh_system_contents))
         return Result.MODIFIED
     else:
         return Result.NO_CHANGE
 
 
-def init_cmd_exe_user(conda_prefix):
+def init_cmd_exe_user(target_path, conda_prefix):
     # HKEY_LOCAL_MACHINE\Software\Microsoft\Command Processor\AutoRun
     # HKEY_CURRENT_USER\Software\Microsoft\Command Processor\AutoRun
     key_str = r'Software\Microsoft\Command Processor'
@@ -656,7 +686,11 @@ def init_cmd_exe_user(conda_prefix):
         new_value = "%s & %s" % (prev_value, hook_path) if prev_value else hook_path
 
         if prev_value != new_value:
-            winreg.SetValueEx(key, "AutoRun", 0, value_type, new_value)
+            if not context.dry_run:
+                winreg.SetValueEx(key, "AutoRun", 0, value_type, new_value)
+            elif context.verbosity:
+                print(target_path)
+                print(make_diff(prev_value, new_value))
             return Result.MODIFIED
         else:
             return Result.NO_CHANGE
@@ -670,11 +704,13 @@ def remove_conda_in_sp_dir(target_path):
     site_packages_dir = target_path
     for fn in glob(join(site_packages_dir, "conda*.egg")):
         print("rm -rf %s" % join(site_packages_dir, fn), file=sys.stderr)
-        rm_rf(join(site_packages_dir, fn))
+        if not context.dry_run:
+            rm_rf(join(site_packages_dir, fn))
         modified = True
     for fn in glob(join(site_packages_dir, "conda.*")):
         print("rm -rf %s" % join(site_packages_dir, fn), file=sys.stderr)
-        rm_rf(join(site_packages_dir, fn))
+        if not context.dry_run:
+            rm_rf(join(site_packages_dir, fn))
         modified = True
     others = (
         "conda",
@@ -685,7 +721,8 @@ def remove_conda_in_sp_dir(target_path):
         path = join(site_packages_dir, other)
         if lexists(path):
             print("rm -rf %s" % path, file=sys.stderr)
-            rm_rf(path)
+            if not context.dry_run:
+                rm_rf(path)
             modified = True
     if modified:
         return Result.MODIFIED
@@ -693,10 +730,10 @@ def remove_conda_in_sp_dir(target_path):
         return Result.NO_CHANGE
 
 
-def make_conda_pth(target_path, conda_source_dir):
+def make_conda_pth(target_path, conda_source_root):
     # target_path: join(site_packages_dir, 'conda-dev.pth')
     conda_pth_path = target_path
-    conda_pth_contents = conda_source_dir
+    conda_pth_contents = conda_source_root
 
     if isfile(conda_pth_path):
         with open(conda_pth_path) as fh:
@@ -705,11 +742,20 @@ def make_conda_pth(target_path, conda_source_dir):
         conda_pth_contents_old = ""
 
     if conda_pth_contents_old != conda_pth_contents:
-        with open(conda_pth_path, 'w') as fh:
-            fh.write(conda_pth_contents)
+        if not context.dry_run:
+            with open(conda_pth_path, 'w') as fh:
+                fh.write(conda_pth_contents)
+        elif context.verbosity:
+            print(target_path)
+            print(make_diff(conda_pth_contents_old, conda_pth_contents), file=sys.stderr)
         return Result.MODIFIED
     else:
         return Result.NO_CHANGE
+
+
+def make_diff(old, new):
+    from difflib import unified_diff
+    return '\n'.join(unified_diff(old.splitlines(), new.splitlines()))
 
 
 if __name__ == "__main__":
